Move getRelativeTimeString/isWithinLast/isWithinNext to utils/datetime
